refactor: 데이터 없을시 0으로 반환하는 코드로 변경

+++ total-repo/server/src/main/java/com/d108/project/domain/area/analytic/dto/IndustryInfoDto.java
@@ -2,9 +2,9 @@ package com.d108.project.domain.area.analytic.dto;
 
 import java.util.List;
 
-public record IndustryInfoDto(List<Integer> industryInfo, List<String> info) {
+public record IndustryInfoDto(List<Long> industryInfo, List<String> info) {
 
-    public IndustryInfoDto(List<Integer> industryInfo, List<String> info) {
+    public IndustryInfoDto(List<Long> industryInfo, List<String> info) {
         this.industryInfo = industryInfo;
         this.info = info;
     }
